scan: mostly eliminate "bundles" as a concept
Before, the internal pipeline passed along object literals containing
metadata (ZIP file info, FLAC stream data, metadata tags, model entities
representing filesystem tracks, albums, and artists, etc). It was all
one flat namespace, with each step accreting new data onto the previous,
until a track had been fully read off the filesystem and its contents
scanned. This lack of organization made the code operating on these
"bundles" hard to follow.

Replace all that with real model entities, so a Track now can have a
sourceArchive that's an Archive associated with it, the metadata from
the ZIP file is associated with its Archive, and the metadata from the
FLAC stream is associated with the AudioFile attached to the Track.

Given the lack of types for FSTrack, FSAlbum, and FSArtist distinct from
their metadata-driven counterparts, and given the fact that metadata
tracks have only string artists and albums, there's still some bundling
going on in the filesystem scanner that assembles the tracks into
albums. This will need to be ironed out to make the auditing work better
anyway, but it's not done yet.