metadata: part 1 of adding support for ID3v2 tags
Take the FLAC scanner and turn it into an abstract metadata scanner that
calls a similar abstract metadata stream reader, which in turn invokes
the specific stream readers for different kinds of audio containers.
Right now, only the FLAC path is tested, because the next step is to
figure out the common vocabulary of tags and replace flacTags everywhere
with `tags` and `musicbrainzTags` objects with standardized keys.