test: FLAC metadata writing had a race
There's a long and convoluted story behind this:

Problem: about 1 out of every 4 test runs, the unpack tests would fail due to
finding the wrong number of albums in a zipfile. The albums are assembled from
metadata, and the extra album would have all the default values from the names,
indicating that the FLAC metadata wasn't being set on the extracted tracks.

Hypothesis: in failing cases, the FLAC reader is getting only part of a chunk
and is terminating early because it gets a finish event before it's finished
looking at chunks. (Or at least this is what I spent several hours thinking.)

          / parserGauge - parser
zipStream
          \ writeGauge - extracted

          / Through - Tokenizr
fd_slicer
          \ Through - fs.WritableStream

- stream events are the same on failing and succeeding reads
- nothing unusual happening in the stream buffering (checked with stream)
- see what's getting written to the FLAC parser by monkeypatching
  `parser.write`:

on success:
002 - 01 FLACParser 8390
002 - 02 FLACParser 8390
005 - 01 FLACParser 8386
005 - 02 FLACParser 8383

on failure:
002 - 01 FLACParser 8282
002 - 02 FLACParser 8390
005 - 01 FLACParser 8386
005 - 02 FLACParser 8282

- from this, it looks like the zipstream isn't sending all the chunks to the
  FLAC parser.
- see if all the chunks are getting written to the output files by
  monkeypatching `extracted.write`:

on success:
002 - 01 extracted 8390 FLACParser 8390
002 - 02 extracted 8390 FLACParser 8390
005 - 01 extracted 8386 FLACParser 8386
005 - 02 extracted 8383 FLACParser 8383

on failure:
002 - 01 extracted 8282 FLACParser 8282
002 - 02 extracted 8390 FLACParser 8390
005 - 01 extracted 8386 FLACParser 8386
005 - 02 extracted 8383 FLACParser 8383

- because there aren't any errors, and the length of the failed files is the
  same as `empty.lac`, it looks like the test code that writes the FLAC tags is
  incorrect, which is both reassuring and super irritating.

CONCLUSION: The reading code was fine, but there was a race condition in the
code that populated stub FLAC tracks with metadata. Once that was fixed, the
problem went away.