zip: refactor to make data dependencies clearer
While debugging the issues found in in the previous commit, I spent a
bunch of time suspecting that the reading (rather than writing) code was
at fault, and changed a bunch of things around to make it clearer what's
waiting on which parts of the flow while extracting entries from the zip
file. I also made the piping code a little more idiomatic. Finally, the
logging levels and the things logged were pretty out of whack, so I
cleaned that up.