zip: depromisify use of yauzl.open
It turns out that the delaty introduced by introducing a microtask wait
when using promisified yauzl.open caused a race that could, when lost,
lead to zipfile entries being emitted before the listener was in place.
Probably a bug, or maybe just another edge case that occurs when jamming
Node streams together with promises.