transcode: tag files with eyeD3
In the process of introducing validation for the transcoded MP3 files
with mp3val, I learned that `flac-metadata` was corrupting the FLAC
stream as part of writing its metadata. I was unable to find a way to
get it to act as a streaming metadata parser without breaking the
resulting files (using both the recommended code and my own best
guesses), so I ripped out `flac-metadata` and did a straight streaming
copy and started running metaflac as a child process. This seems to be a
theme on this project.

After doing that, I introduced basic tag writing using eyeD3. eyeD3 can
handle all the frame types I want to use, but unlike metaflac, it
doesn't have any way to read tag / value pairs from a file or stdin, so
the command lines get very long. This work has a ways left to go, but
it currently handles basic metadata and produces valid MP3s.

I also switched to using `mississippi` for streams handling for peace of
mind, which is already making the streaming code clearer. Thanks,
maxogden! Thanks, mafintosh!